It is the Desktop java Application
In this software the Admin can add the category in his menu. He can add new Items with its photo. Admin can modify the records and can export his total bills detail in Excel format.
On the other hand the manager will fill the details for the billing. After the final checkout the reciept will be printed and will be send to the customer via EMAIL.